+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## v3.0.63 - 2026-09-08
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.62...v3.0.63)
+
+- [#7685](https://github.com/ORCID/ORCID-Source/pull/7685): PD-6145 per element source
+
+## v3.0.62 - 2026-09-08
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.61...v3.0.62)
+
+- [#7684](https://github.com/ORCID/ORCID-Source/pull/7684): Pd-6145 More serialization fixes
+
+## v3.0.61 - 2026-09-08
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.60...v3.0.61)
+
+- [#7682](https://github.com/ORCID/ORCID-Source/pull/7682): Fixed the cache for identifier types
+
 ## v3.0.60 - 2026-09-08
 
 [Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.59...v3.0.60)